# pkg/api/v1/docs/api_swagger.yaml
basePath: /v1
consumes:
- application/json
definitions:
  models.APIKeyResponse:
    properties:
      authority_name:
        type: string
      client_template_name:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      secret:
        type: string
      server_template_name:
        type: string
    type: object
...
  models.CreateAPIKeyRequest:
    properties:
      authority_name:
        type: string
      client_template_name:
        type: string
      name:
        type: string
      server_template_name:
        type: string
    type: object
...
  models.DeleteAPIKeyRequest:
    properties:
      authority_name:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: admin@loopholelabs.io
    name: API Support
  description: Returns the health and status of the various services that make up
    the API.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://loopholelabs.io/privacy
  title: EndKey API V1
  version: "1.0"
paths:
  /apikey:
    delete:
      consumes:
      - application/json
      description: Delete an API Key
      parameters:
      - description: Delete API Key Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DeleteAPIKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "412":
          description: Precondition Failed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      tags:
      - apikey
    post:
      consumes:
      - application/json
      description: Create a new API Key for a given Authority
      parameters:
      - description: Create API Key Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIKeyResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "412":
          description: Precondition Failed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      tags:
      - apikey
  /apikey/{authority_name}:
    get:
      consumes:
      - application/json
      description: Lists all the api keys
      parameters:
      - description: Authority Name
        in: path
        name: authority_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.APIKeyResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "412":
          description: Precondition Failed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      tags:
      - apikey
...
produces:
- application/json
schemes:
- https
swagger: "2.0"