# pkg/api/v1/models/apikey.go
/*
	Copyright 2023 Loophole Labs

	Licensed under the Apache License, Version 2.0 (the "License");
	you may not use this file except in compliance with the License.
	You may obtain a copy of the License at

		   http://www.apache.org/licenses/LICENSE-2.0

	Unless required by applicable law or agreed to in writing, software
	distributed under the License is distributed on an "AS IS" BASIS,
	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
	See the License for the specific language governing permissions and
	limitations under the License.
*/

package models

type CreateAPIKeyRequest struct {
	Name               string `json:"name"`
	AuthorityName      string `json:"authority_name"`
	ServerTemplateName string `json:"server_template_name"`
	ClientTemplateName string `json:"client_template_name"`
}

type DeleteAPIKeyRequest struct {
	Name          string `json:"name"`
	AuthorityName string `json:"authority_name"`
}

type APIKeyResponse struct {
	CreatedAt          string `json:"created_at"`
	ID                 string `json:"id"`
	Name               string `json:"name"`
	AuthorityName      string `json:"authority_name"`
	ServerTemplateName string `json:"server_template_name"`
	ClientTemplateName string `json:"client_template_name"`
	Secret             string `json:"secret"`
}